Use a Kaplan-Meier estimator to estimate IPCW weights

Computes inverse probability of censoring weights (IPCW) for right-censored survival data using the pec::ipcw function.

Usage

ipcw_weights(data, tau, time_var = "t", status_var = "delta")

Arguments

data

A data frame containing the survival data. Must include columns for the observed time and event indicator.

tau

Numeric scalar. The time horizon at which the survival probability is to be estimated.

time_var

Character. The name of the variable in data representing the observed time to event or censoring. Default is "t".

status_var

Character. The name of the variable in data representing the event indicator (1 if event occurred, 0 if censored). Default is "delta".

Value

A numeric vector of IPCW weights, ordered as in the original data. The weights are not normalized (are not scaled to sum to one).

Details

Predicting a survival probability for a right-censored event time outcome at a specific time horizon \(\tau\) can be performed using binary classifiers. For this purpose, the event time outcome is dichotomized into a binary variable indicating whether a subject has survived up to time \(\tau\) or not. However, this approach can not account for subjects who are lost to follow-up before time \(\tau\) and therefore before experiencing the event of interest. This leads to worse discrimination and calibration of the model (Reps et al. 2021; Kvamme and Borgan 2023) .

Weighting observations by their Inverse Probability of Censoring has been proposed to prevent bias (Vock et al. 2016; Ginestet et al. 2021; Blanche et al. 2023) and improve model performance.

\[ \tilde{w}_i = \left\lbrace \begin{array}{ll} 0, & c_i < \tau \wedge t^*_i \cr \frac{1}{\mathbb{P}(C > \tau \mid X = x_i)}, & \tau < c_i \wedge t^*_i \cr \frac{1}{\mathbb{P}(C > t_i \mid X = x_i)}, & t^*_i < c_i \wedge \tau \end{array} \right. \]

The function uses the pec::ipcw(Gerds 2023) function to compute the IPCW weights. They are not normalized (are not scaled to sum to one).

References

Blanche PF, Holt A, Scheike T (2023). “On logistic regression with right censored data, with or without competing risks, and its use for estimating treatment effects.” Lifetime Data Analysis, 29(2), 441–482. ISSN 1380-7870, doi:10.1007/s10985-022-09564-6 .

Gerds TA (2023). pec: Prediction Error Curves for Risk Prediction Models in Survival Analysis. R package version 2023.04.12, https://CRAN.R-project.org/package=pec.

Ginestet GP, Kotalik A, Vock DM, Wolfson J, Gabriel EE (2021). “Stacked Inverse Probability of Censoring Weighted Bagging A Case Study In the InfCareHIV Register.” Journal of the Royal Statistical Society Series C: Applied Statistics, 70, 51–65. doi:10.1111/rssc.12448 , https://academic.oup.com/jrsssc/article/70/1/51/7033916 http://dx.doi.org/10.1111/rssc.12448.

Kvamme H, Borgan Ø (2023). “The Brier Score under Administrative Censoring: Problems and Solutions.” Journal of Machine Learning Research, 24, 1–26. 1912.08581, http://arxiv.org/abs/1912.08581.

Reps JM, Rijnbeek P, Cuthbert A, Ryan PB, Pratt N, Schuemie M (2021). “An empirical analysis of dealing with patients who are lost to follow-up when developing prognostic models using a cohort design.” BMC medical informatics and decision making, 21(1), 43. ISSN 1472-6947, doi:10.1186/s12911-021-01408-x , http://www.ncbi.nlm.nih.gov/pubmed/33549087.

Vock DM, Wolfson J, Bandyopadhyay S, Adomavicius G, Johnson PE, Vazquez-Benitez G, O'Connor PJ (2016). “Adapting machine learning techniques to censored time-to-event health record data A general-purpose approach using inverse probability of censoring weighting.” Journal of Biomedical Informatics, 61, 119–131. doi:10.1016/j.jbi.2016.03.009 , https://doi.org/10.1016/j.jbi.2016.03.009.

See also

Examples

data <- data.frame(
  t = c(5, 8, 12, 15, 20),
  delta = c(1, 0, 1, 0, 1)
)
w <- ipcw_weights(data, tau = 10)