IPCW XGBoost Binary Classifier

Fits a binary classification model using XGBoost with IPCW for right-censored survival data. Hyperparameter tuning and Jackknife model training are performed.

Usage

ipcw_xgboost(
  data,
  tau,
  time_var = "t",
  status_var = "delta",
  verbose = 0,
  grid = ipcw_xgboost_default_grid(),
  nrounds = 100,
  early_stopping_rounds = 10,
  nfold = 3,
  nthread = 1
)

ipcw_xgboost_default_grid()

Arguments

data

A data frame containing the survival data. Must include columns for the observed time and event indicator.

tau

Numeric scalar. The time horizon at which the survival probability is to be estimated.

time_var

Character. The name of the variable in data representing the observed time to event or censoring. Default is "t".

status_var

Character. The name of the variable in data representing the event indicator (1 if event occurred, 0 if censored). Default is "delta".

verbose

Integer. Verbosity level for XGBoost training and cross-validation (default is 0).

grid

Data frame. Grid of hyperparameters to test in cross-validation. The default is the output of ipcw_xgboost_default_grid().

nrounds

Integer. Maximum number of boosting rounds for XGBoost training and cross-validation (default is 100).

early_stopping_rounds

Integer. Number of rounds with no improvement to trigger early stopping during cross-validation (default is 10).

nfold

Integer. Number of folds for cross-validation (default is 3).

nthread

Integer. Number of threads to use for XGBoost training (default is 1).

Value

An object of class IPCWModel.

Details

Training is performed using the xgboost package (Chen and Guestrin 2016) . The weights are calculated using the

Functions

ipcw_xgboost_default_grid(): Returns a default grid of hyperparameters.

References

Chen T, Guestrin C (2016). “XGBoost: A Scalable Tree Boosting System.” In Proceedings of the 22nd ACM SIGKDD International Conference on Knowledge Discovery and Data Mining, KDD '16, 785–794. ISBN 978-1-4503-4232-2, doi:10.1145/2939672.2939785 , http://doi.acm.org/10.1145/2939672.2939785.

See also

Examples

#' # veteran data example
library(survival)
tau <- 80
df <- veteran[, c("time", "status", "trt")]
newdata <- data.frame(trt = c(1, 2))

fit <- ipcw_xgboost(df,
  tau = tau, time_var = "time",
  status_var = "status"
)
predict(fit, newdata)
#>   prediction     lower     upper           se
#> 1  0.4999998 0.4999986 0.5000009 5.848756e-07
#> 2  0.4999996 0.4999976 0.5000017 1.033711e-06